리팩토링 : 리액트 애플리케이션 컴포넌트 및 메서드명 수정
