+++ spring-app/src/main/java/kr/co/co_working/memberWorkspace/repository/MemberWorkspaceRepository.java
@@ -7,8 +7,5 @@ import java.util.List;
 
 public interface MemberWorkspaceRepository extends JpaRepository<MemberWorkspace, Long> {
     MemberWorkspace findByMemberEmailAndWorkspaceId(String email, Long workspaceId);
-
-    List<MemberWorkspace> findByMemberEmail(String email);
-
     List<MemberWorkspace> findByWorkspaceId(Long workspaceId);
 }
